backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
redis
rq
faster-whisper
ctranslate2
openai-whisper